+++ src/types/navigator.ts
@@ -6,7 +6,7 @@ export type BottomTabScreenName = keyof BottomTabParamList;
 
 export type RootStackParamList = {
   BottomTabScreens: undefined;
-  // LoginScreen: undefined;
+  LoginScreen: undefined;
   //TODO: 추후에 필요한 페이지 타입 주가
 };
